backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.6
djangorestframework==3.14.0
finnhub-python==2.4.19
gunicorn==20.1.0
idna==3.4
psycopg==3.1.12
psycopg-binary==3.1.12
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7

